fix: split EngineServer into domain-specific handler files (#200)
